fix compiler warnings
git-svn-id: svn://anonsvn.kde.org/home/kde/trunk/kdemultimedia/kmidi@31387 283d02a7-25f6-0310-bc7c-ecb5cbfe19da

+++ kmidi.cpp
@@ -835,7 +835,7 @@ void KMidi::setPatch( int index )
 
 void KMidi::setSong( int number )
 {
-    if (!playlist->count() || number > playlist->count()) {
+    if (!playlist->count() || number > (int)playlist->count()) {
 	redoplaybox();
 	return;
     }
